Moving all css files to folder style
