+++ src/components/Menu.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./Menu.css";
+import "../style/Menu.css";
 
 const Menu = () => {
   return (
