+++ src/index.js
@@ -4,7 +4,7 @@ import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 import rootReducer from './reducers/index';
 import App from './components/App';
-import './index.css';
+import './style/index.css';
 
 
 const initialState = {
